moved downloading fim.so from setup.py to spade
