+++ elephant/spade.py
@@ -66,7 +66,13 @@ import time
 import quantities as pq
 import warnings
 from elephant.spade_src import fast_fca
+from elephant.spade_src.fim_manager import download_spade_fim
+
 warnings.simplefilter('once', UserWarning)
+
+download_spade_fim()
+
+
 try:
     from mpi4py import MPI  # for parallelized routines
     HAVE_MPI = True
