+++ setup.py
@@ -1,48 +1,11 @@
 # -*- coding: utf-8 -*-
 
 import os
-import platform
-import struct
 import sys
 
 from setuptools import setup
-from setuptools.command.install import install
 
 python_version_major = sys.version_info.major
-
-if python_version_major == 2:
-    from urllib import urlretrieve
-else:
-    from urllib.request import urlretrieve
-
-
-def download_spade_fim():
-    """
-    Downloads SPADE specific PyFIM binary file.
-    """
-    arch_bits = struct.calcsize("P") * 8
-    if platform.system() == "Windows":
-        fim_filename = "fim.pyd"
-    else:
-        # Linux
-        fim_filename = "fim.so"
-    url_fim = "http://www.borgelt.net/bin{arch}/py{py_ver}/{filename}". \
-        format(arch=arch_bits, py_ver=python_version_major,
-               filename=fim_filename)
-
-    try:
-        urlretrieve(url_fim,
-                    filename=os.path.join('elephant', 'spade_src',
-                                          fim_filename))
-    except:
-        print("Unable to download {url} module.".format(url=url_fim))
-
-
-class ElephantInstall(install):
-    def run(self):
-        install.run(self)
-        download_spade_fim()
-
 
 with open("README.rst") as f:
     long_description = f.read()
@@ -52,7 +15,6 @@ extras_require = {}
 for extra in ['extras', 'docs', 'tests']:
     with open('requirements-{0}.txt'.format(extra)) as fp:
         extras_require[extra] = fp.read()
-
 
 setup(
     name="elephant",
@@ -68,7 +30,6 @@ setup(
         os.path.join('spade_src', '*.so'),
         os.path.join('spade_src', '*.pyd'),
     ]},
-    cmdclass={'install': ElephantInstall},
 
     install_requires=install_requires,
     extras_require=extras_require,
